+++ lib/src/nodes/expressions/call.dart
@@ -1,3 +1,5 @@
+import 'dart:mirrors';
+
 import '../core.dart';
 
 class Call extends Expression {
@@ -44,7 +46,7 @@ class Call extends Expression {
       }
     }
 
-    return Function.apply(expr.resolve(context).call as Function, args, kwargs);
+    return reflect(expr.resolve(context)).invoke(#call, args, kwargs).reflectee;
   }
 
   @override